move entity stuff to file
