+++ src/lib/vocab/file/file.ts
@@ -1,3 +1,5 @@
+import {GUEST_PERSONA_NAME} from '$lib/vocab/persona/constants';
+
 export interface File {
 	file_id: number;
 	actor_id: number;
@@ -20,3 +22,12 @@ export const FileSchema = {
 	required: ['file_id', 'actor_id', 'space_id', 'content', 'created', 'updated'],
 	additionalProperties: false,
 };
+
+// TODO expand to the entire vocabulary? generate if so
+export type EntityType = 'Persona' | 'Community';
+
+export const toName = (entity: null | undefined | {name?: string}): string =>
+	(entity as any)?.name ?? GUEST_PERSONA_NAME;
+
+export const toIcon = (entity: null | undefined | {icon?: string}): string | null =>
+	(entity as any)?.icon ?? null;
